+++ backintime-debugger/data/src/commonMain/kotlin/com/github/kitakkun/backintime/debugger/data/driver/SqlDriver.kt
@@ -1,4 +1,4 @@
-package com.github.kitakkun.backintime.debugger.data
+package com.github.kitakkun.backintime.debugger.data.driver
 
 import app.cash.sqldelight.db.SqlDriver
